# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Inngest App API",
    "version": "1.0.0",
    "description": "API documentation for the Inngest app"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your bearer token in the format: Bearer {token}"
      }
    },
    "schemas": {
...
      "WorkflowInputField": {
        "type": "object",
        "required": [
          "key",
          "label",
          "type"
        ],
        "properties": {
          "key": {
            "type": "string",
            "minLength": 1
          },
          "label": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "text",
              "number",
              "boolean",
              "json",
              "credential",
              "email",
              "url",
              "date",
              "time",
              "file",
              "select",
              "multiselect"
            ]
          },
          "required": {
            "type": "boolean",
            "default": false
          },
          "defaultValue": {},
          "validation": {
            "type": "object",
            "properties": {
              "min": {
                "type": "number"
              },
              "max": {
                "type": "number"
              },
              "pattern": {
                "type": "string"
              },
              "options": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "WorkflowTemplateResponse": {
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "category",
          "eventName"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": [
              "REPORTING",
              "DATA_PROCESSING",
              "COMMUNICATION",
              "AUTOMATION",
              "ANALYTICS",
              "INTEGRATION",
              "MONITORING",
              "CUSTOM"
            ]
          },
          "version": {
            "type": "string",
            "default": "1.0.0"
          },
          "eventName": {
            "type": "string"
          },
          "canBeScheduled": {
            "type": "boolean",
            "default": false
          },
          "requiredProviders": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "GOOGLE",
                "SLACK",
                "HUBSPOT",
                "FIRECRAWL",
                "CUSTOM"
              ]
            }
          },
          "inputFields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkflowInputField"
            }
          },
          "outputSchema": {
            "type": "object"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "WorkflowCreateRequest": {
        "type": "object",
        "required": [
          "name",
          "description",
          "workflowDefinitionId"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          },
          "category": {
            "type": "string",
            "enum": [
              "REPORTING",
              "DATA_PROCESSING",
              "COMMUNICATION",
              "AUTOMATION",
              "ANALYTICS",
              "INTEGRATION",
              "MONITORING",
              "CUSTOM"
            ]
          },
          "workflowDefinitionId": {
            "type": "string",
            "minLength": 1
          },
          "enabled": {
            "type": "boolean",
            "default": true,
            "description": "Whether the workflow instance is enabled for use by the user"
          },
          "isActive": {
            "type": "boolean",
            "default": false,
            "description": "Whether the workflow is currently active/scheduled to run"
          },
          "canBeScheduled": {
            "type": "boolean",
            "default": false
          },
          "cronExpressions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "default": []
          },
          "timezone": {
            "type": "string",
            "default": "UTC"
          },
          "input": {
            "type": "object"
          },
          "credentialIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "default": []
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "default": []
          }
        }
      },
...
      "WorkflowResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "type": "string",
            "enum": [
              "REPORTING",
              "DATA_PROCESSING",
              "COMMUNICATION",
              "AUTOMATION",
              "ANALYTICS",
              "INTEGRATION",
              "MONITORING",
              "CUSTOM"
            ]
          },
          "workflowDefinitionId": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean",
            "description": "Whether the workflow instance is enabled for use by the user"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the workflow is currently active/scheduled to run"
          },
          "canBeScheduled": {
            "type": "boolean"
          },
          "cronExpressions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "timezone": {
            "type": "string",
            "nullable": true
          },
          "lastRunAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "nextRunAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "input": {
            "type": "object",
            "nullable": true
          },
          "eventName": {
            "type": "string"
          },
          "requiredProviders": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "GOOGLE",
                "SLACK",
                "HUBSPOT",
                "FIRECRAWL",
                "CUSTOM"
              ]
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "credentials": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "type": {
                  "type": "string"
                },
                "provider": {
                  "type": "string"
                },
                "config": {
                  "type": "object",
                  "nullable": true
                }
              }
            }
          },
          "template": {
            "$ref": "#/components/schemas/WorkflowTemplateResponse"
          }
        }
      },
...
      "PaginationResponse": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "minimum": 1
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "totalPages": {
            "type": "integer",
            "minimum": 0
          },
          "hasNext": {
            "type": "boolean"
          },
          "hasPrev": {
            "type": "boolean"
          }
        }
      },
...
    },
...
  },
  "paths": {
...
    "/api/workflows/install/{templateId}": {
      "post": {
        "summary": "Install a workflow template",
        "description": "Creates a new workflow instance from a marketplace template for the authenticated user.",
        "tags": [
          "Workflows"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "templateId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The workflow template ID to install"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Custom name for this workflow instance"
                  },
                  "input": {
                    "type": "object",
                    "description": "Initial configuration for the workflow"
                  },
                  "credentialIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Credential IDs to associate with this workflow"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Workflow template installed successfully"
          },
          "400": {
            "description": "Invalid template or validation failed"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Template not found or user not found"
          },
          "409": {
            "description": "Template already installed"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/workflows": {
      "get": {
        "summary": "Get user's installed workflows",
        "description": "Returns a paginated list of workflow instances that the user has installed/configured, including default workflows and marketplace workflows they've added.",
        "tags": [
          "Workflows"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "description": "Number of items per page"
          },
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string",
              "enum": [
                "REPORTING",
                "DATA_PROCESSING",
                "COMMUNICATION",
                "AUTOMATION",
                "ANALYTICS",
                "INTEGRATION",
                "MONITORING",
                "CUSTOM"
              ]
            },
            "description": "Filter by workflow category"
          },
          {
            "in": "query",
            "name": "enabled",
            "schema": {
              "type": "boolean"
            },
            "description": "Filter by enabled status"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Search workflows by name or description"
          },
          {
            "in": "query",
            "name": "tags",
            "schema": {
              "type": "string"
            },
            "description": "Filter by tags (comma-separated)"
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "enum": [
                "name",
                "createdAt",
                "updatedAt",
                "lastRunAt"
              ],
              "default": "createdAt"
            },
            "description": "Field to sort by"
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "desc"
            },
            "description": "Sort order"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of workflows",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/WorkflowResponse"
                          }
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/PaginationResponse"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Create a new workflow",
        "description": "Creates a new workflow instance from a workflow definition with validated input.",
        "tags": [
          "Workflows"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Workflow created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/WorkflowResponse"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data or workflow input validation failed"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Workflow definition not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
...
}